- Clean up package structure - Simplify class set by removing DTOs - Fix AMQP connection by using Docker host name - Rename classes to include Production prefix - Remove example code - Extend .gitignore files
