+++ src/systems/assembly/src/main/java/dk/sdu/se23g6/arch/assembly/MainApplication.java
@@ -1,11 +1,10 @@
-package dk.sdu.se23g6.arch.projecttitle;
+package dk.sdu.se23g6.arch.assembly;
 
 
-import dk.sdu.se23g6.arch.projecttitle.models.order.Order;
-import dk.sdu.se23g6.arch.projecttitle.models.order.OrderStep;
-import dk.sdu.se23g6.arch.projecttitle.models.order.StepStatus;
-import dk.sdu.se23g6.arch.projecttitle.models.order.dto.OrderStepDTO;
-import org.springframework.amqp.core.Message;
+import dk.sdu.se23g6.arch.assembly.model.ProductionOrder;
+import dk.sdu.se23g6.arch.assembly.model.ProductionStep;
+import dk.sdu.se23g6.arch.assembly.model.StepStatus;
+import org.springframework.amqp.core.MessageProperties;
 import org.springframework.amqp.core.Queue;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
@@ -14,7 +13,6 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.messaging.handler.annotation.Payload;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.scheduling.annotation.EnableAsync;
 
 import java.util.Random;
@@ -57,23 +55,22 @@ public class MainApplication {
   }
 
   @RabbitListener(queues = RECEIVER_QUEUE)
-  public void listen(@Payload Order in) {
+  public void listen(@Payload ProductionOrder in) {
     System.out.println("Received order from the Supervisor system with orderId " + in.getOrderId());
 
     Random gen = new Random();
     System.out.println("Processing " + in.getSteps().size() + " steps.");
 
-    for (OrderStep step : in.getSteps()) {
-      boolean shouldComplete = gen.nextBoolean();
+    for (ProductionStep step : in.getSteps()) {
       try {
         Thread.sleep(3000);
       } catch (InterruptedException e) {
         System.out.println("Timeout error happened");
       }
       step.setOrderStatus(StepStatus.COMPLETED);
-      System.out.println("Completing orders " + );
-      OrderStepDTO orderStepDTO = new OrderStepDTO(in.getOrderId(), step.getStepId(), StepStatus.COMPLETED);
-      this.template.send(SENDER_QUEUE, converter.toMessage(orderStepDTO, null));
+      System.out.println("Completing orders ");
+      ProductionStep productionStepDTO = new ProductionStep(in.getOrderId(), step.getStepId(), StepStatus.COMPLETED);
+      this.template.send(SENDER_QUEUE, converter.toMessage(productionStepDTO, new MessageProperties()));
     }
   }
