+++ src/systems/assembly/src/main/java/dk/sdu/se23g6/arch/assembly/model/ProductionOrderResponse.java
@@ -1,0 +1,7 @@
+package dk.sdu.se23g6.arch.assembly.model;
+
+
+public class ProductionOrderResponse {
+    private int robotId;
+    private int orderId;
+}
