+++ src/systems/assembly/src/main/java/dk/sdu/se23g6/arch/assembly/model/StepStatus.java
@@ -1,4 +1,4 @@
-package dk.sdu.se23g6.arch.projecttitle.models.order;
+package dk.sdu.se23g6.arch.assembly.model;
 
 public enum StepStatus {
     NOT_STARTED,
