+++ src/systems/assembly/src/main/java/dk/sdu/se23g6/arch/assembly/production/ProductionOrderHandler.java
@@ -1,0 +1,4 @@
+package dk.sdu.se23g6.arch.assembly.production;
+
+public class ProductionOrderHandler {
+}
