--- src/systems/assembly/src/main/java/dk/sdu/se23g6/arch/projecttitle/example/RandomDataGenerator.java
@@ -1,51 +1,0 @@
-package dk.sdu.se23g6.arch.projecttitle.example;
-
-import dk.sdu.se23g6.arch.projecttitle.models.assemblysystem.AssemblyOrder;
-
-import java.util.UUID;
-import java.time.Instant;
-import java.security.SecureRandom;
-import java.util.Random;
-
-public class RandomDataGenerator {
-    private static SecureRandom secureRandom = new SecureRandom();
-    private static Random random = new Random();
-
-    // Generate a random UUID as an Id
-    public static String generateRandomId() {
-        return UUID.randomUUID().toString();
-    }
-
-    // Generate a random timestamp
-    public static long generateRandomTimestamp() {
-        return Instant.now().toEpochMilli();
-    }
-
-    // Generate a random string from a predefined list
-    public static String generateRandomStatus() {
-        String[] statusList = {"running", "error", "Idle", "waiting"};
-        int randomIndex = random.nextInt(statusList.length);
-        return statusList[randomIndex];
-    }
-
-    public static RandomData generateRandomData() {
-        return RandomData.builder()
-                .withStepId(generateRandomId())
-                .withOrderId(generateRandomId())
-                .withRobotId(generateRandomId())
-                .withTimestamp(generateRandomTimestamp())
-                .withStatus(generateRandomStatus())
-                .build();
-    }
-
-    public AssemblyOrder createRandomResponse(String id) {
-        return new AssemblyOrder(
-                generateRandomId(),
-                id,
-                generateRandomId(),
-                generateRandomTimestamp(),
-                generateRandomStatus()
-        );
-    }
-
-}
