--- src/systems/assembly/src/main/java/dk/sdu/se23g6/arch/projecttitle/models/order/OrderError.java
@@ -1,4 +1,0 @@
-package dk.sdu.se23g6.arch.projecttitle.models.order;
-
-public record OrderError(Order order, String message) {
-}
