--- src/systems/assembly/src/main/java/dk/sdu/se23g6/arch/projecttitle/models/order/OrderResponse.java
@@ -1,7 +1,0 @@
-package dk.sdu.se23g6.arch.projecttitle.models.order;
-
-
-public class OrderResponse {
-    private int robotId;
-    private int orderId;
-}
