--- src/systems/assembly/src/main/java/dk/sdu/se23g6/arch/projecttitle/models/order/dto/CreateOrderDTO.java
@@ -1,6 +1,0 @@
-package dk.sdu.se23g6.arch.projecttitle.models.order.dto;
-
-import java.util.List;
-
-public record CreateOrderDTO(List<String> steps) {
-}
