--- src/systems/assembly/src/main/java/dk/sdu/se23g6/arch/projecttitle/models/order/dto/OrderStepDTO.java
@@ -1,16 +1,0 @@
-package dk.sdu.se23g6.arch.projecttitle.models.order.dto;
-
-import dk.sdu.se23g6.arch.projecttitle.models.order.StepStatus;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@RequiredArgsConstructor
-public class OrderStepDTO {
-    private String orderId;
-    private String stepId;
-    private StepStatus status;
-}
