+++ src/systems/assembly/src/main/resources/application.properties
@@ -5,7 +5,7 @@ server.servlet.contextPath=${SPRING_BASE_PATH:}
 server.error.include-message=always
 
 # RabbitMQ - AMQP
-spring.rabbitmq.host=localhost
+spring.rabbitmq.host=rabbitmq # Because the docker host is called rabbitmq.
 spring.rabbitmq.port=5672
 spring.rabbitmq.username=admin
 spring.rabbitmq.password=123
